+++ src/main/java/me/marin/lockout/lockout/DefaultGoalRegister.java
@@ -19,6 +19,7 @@ import me.marin.lockout.lockout.goals.dimension.EnterEndGoal;
 import me.marin.lockout.lockout.goals.dimension.EnterNetherGoal;
 import me.marin.lockout.lockout.goals.experience.ReachXPLevel15Goal;
 import me.marin.lockout.lockout.goals.experience.ReachXPLevel30Goal;
+import me.marin.lockout.lockout.goals.have_more.HaveMostAdvancementsGoal;
 import me.marin.lockout.lockout.goals.have_more.HaveMostUniqueCraftsGoal;
 import me.marin.lockout.lockout.goals.have_more.HaveMostXPLevelsGoal;
 import me.marin.lockout.lockout.goals.kill.*;
@@ -37,6 +38,9 @@ import me.marin.lockout.lockout.goals.status_effect.*;
 import me.marin.lockout.lockout.goals.status_effect.unique.Get3StatusEffectsGoal;
 import me.marin.lockout.lockout.goals.status_effect.unique.Get4StatusEffectsGoal;
 import me.marin.lockout.lockout.goals.status_effect.unique.Get6StatusEffectsGoal;
+import me.marin.lockout.lockout.goals.status_effect.applied_for_x_minutes.HaveEffectsAppliedFor5MinutesGoal;
+import me.marin.lockout.lockout.goals.status_effect.applied_for_x_minutes.HaveEffectsAppliedFor8MinutesGoal;
+import me.marin.lockout.lockout.goals.status_effect.applied_for_x_minutes.HaveEffectsAppliedFor10MinutesGoal;
 import me.marin.lockout.lockout.goals.tame_animal.TameCatGoal;
 import me.marin.lockout.lockout.goals.tame_animal.TameHorseGoal;
 import me.marin.lockout.lockout.goals.tame_animal.TameParrotGoal;
@@ -498,6 +502,16 @@ public class DefaultGoalRegister {
         INSTANCE.register(GoalType.OBTAIN_5_UNIQUE_STONE_TYPES, Obtain5UniqueStoneGoal.class);
 
         INSTANCE.register(GoalType.DIE_TO_WARDEN, DieToWardenGoal.class);
+
+        INSTANCE.register(GoalType.HAVE_EFFECTS_APPLIED_FOR_5_MINUTES, HaveEffectsAppliedFor5MinutesGoal.class);
+        INSTANCE.register(GoalType.HAVE_EFFECTS_APPLIED_FOR_8_MINUTES, HaveEffectsAppliedFor8MinutesGoal.class);
+        INSTANCE.register(GoalType.HAVE_EFFECTS_APPLIED_FOR_10_MINUTES, HaveEffectsAppliedFor10MinutesGoal.class);
+        INSTANCE.register(GoalType.HAVE_MOST_ADVANCEMENTS, HaveMostAdvancementsGoal.class, GoalRequirements.TEAMS_GOAL);
+        INSTANCE.register(GoalType.KILL_BLAZE_WITH_SNOWBALL, KillBlazeWithSnowballGoal.class);
+        INSTANCE.register(GoalType.DIE_TO_PUFFERFISH, DieToPufferfishGoal.class, new GoalRequirements.Builder()
+                .biomes(List.of(LUKEWARM_OCEAN, WARM_OCEAN, DEEP_LUKEWARM_OCEAN))
+                .build()
+        );
 
     }